# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DiversityMetadata": {
        "description": "What the search-side Diversity ranker actually did.",
        "properties": {
          "applied": {
            "default": false,
            "title": "Applied",
            "type": "boolean"
          },
          "candidate_count": {
            "default": 0,
            "title": "Candidate Count",
            "type": "integer"
          },
          "depth": {
            "default": "auto",
            "description": "Requested candidate-pool depth: auto, 500, 1000, 2000, 5000, or 10000.",
            "title": "Depth",
            "type": "string"
          },
          "duplicate_images_collapsed": {
            "default": 0,
            "title": "Duplicate Images Collapsed",
            "type": "integer"
          },
          "mode": {
            "default": "off",
            "title": "Mode",
            "type": "string"
          },
          "pool_depth": {
            "default": 0,
            "description": "Number of candidates actually retrieved for the ranking pass.",
            "title": "Pool Depth",
            "type": "integer"
          },
          "requested": {
            "default": false,
            "title": "Requested",
            "type": "boolean"
          },
          "result_count": {
            "default": 0,
            "title": "Result Count",
            "type": "integer"
          },
          "semantic_groups_covered": {
            "default": 0,
            "title": "Semantic Groups Covered",
            "type": "integer"
          },
          "strength": {
            "default": 0.0,
            "title": "Strength",
            "type": "number"
          }
        },
        "title": "DiversityMetadata",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "centroid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "First active centroid name, when any centroids are in play. Kept for backward compat with single-centroid clients; the full list lives in `centroids`. Mutually exclusive with q/positives/negatives.",
            "title": "Centroid"
          },
          "centroids": {
            "description": "Active centroid names in blend order. Empty when no centroid search is in play. One or more names blends via weighted mean.",
            "items": {
              "type": "string"
            },
            "title": "Centroids",
            "type": "array"
          },
          "diverse": {
            "default": false,
            "description": "Backwards-compatible flag; true when search Diversity was applied.",
            "title": "Diverse",
            "type": "boolean"
          },
          "diversity": {
            "$ref": "#/components/schemas/DiversityMetadata",
            "description": "Diagnostics for the search-only Diversity ranking pass."
          },
          "has_more": {
            "default": false,
            "description": "True when more results likely exist on a subsequent page",
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "description": "Max results requested for this page",
            "title": "Limit",
            "type": "integer"
          },
          "negatives": {
            "items": {
              "type": "string"
            },
            "title": "Negatives",
            "type": "array"
          },
          "offset": {
            "default": 0,
            "description": "Offset of this page in the full result set",
            "title": "Offset",
            "type": "integer"
          },
          "positives": {
            "items": {
              "type": "string"
            },
            "title": "Positives",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "title": "Results",
            "type": "array"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque session id. For /api/random this identifies the shuffled deck; pass it back with an incremented `offset` to walk forward. None for non-session endpoints (search, similar, etc.) \u2014 those use offset/limit directly against Qdrant.",
            "title": "Session Id"
          },
          "session_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Total photos in the session deck. Only set for /api/random. Use this with `offset` to know when you've walked everything.",
            "title": "Session Total"
          },
          "surprise": {
            "default": false,
            "description": "True when results were randomly sampled from a deep pool (Surprise Me mode).",
            "title": "Surprise",
            "type": "boolean"
          },
          "took_ms": {
            "title": "Took Ms",
            "type": "integer"
          },
          "view": {
            "default": "grid",
            "description": "Result view requested: 'grid' (default) or 'feed' (single-column, full-width).",
            "title": "View",
            "type": "string"
          },
          "weights": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-centroid weights, same order as `centroids`. None means all weights equal 1.0 (the default).",
            "title": "Weights"
          }
        },
        "required": [
          "query",
          "results",
          "took_ms",
          "limit"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SearchResult": {
        "properties": {
          "blurhash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "LQIP (low-quality image placeholder). Decoded client-side into a tinted background while the real image loads. None when the encoder failed or the point was indexed before the blurhash feature shipped.",
            "title": "Blurhash"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Photo height in pixels, when known. None for older rows that were indexed before width/height were recorded.",
            "title": "Height"
          },
          "id": {
            "description": "Qdrant point id (32-char hex prefix)",
            "title": "Id",
            "type": "string"
          },
          "is_disliked": {
            "default": false,
            "description": "True when the image is marked as a dislike (hides it from future recommendations)",
            "title": "Is Disliked",
            "type": "boolean"
          },
          "is_favorite": {
            "default": false,
            "description": "True when the image is marked as a favourite",
            "title": "Is Favorite",
            "type": "boolean"
          },
          "path": {
            "description": "Absolute source path on the NAS",
            "title": "Path",
            "type": "string"
          },
          "score": {
            "description": "Cosine similarity in [-1, 1]",
            "title": "Score",
            "type": "number"
          },
          "score_str": {
            "default": "",
            "description": "Score formatted to 3 decimals, e.g. '0.873'. Computed server-side so SSR + JS render identically.",
            "title": "Score Str",
            "type": "string"
          },
          "url": {
            "default": "",
            "description": "Public URL for the /photo/{id}/raw endpoint",
            "title": "Url",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Photo width in pixels, when known. None for older rows that were indexed before width/height were recorded.",
            "title": "Width"
          }
        },
        "required": [
          "id",
          "path",
          "score"
        ],
        "title": "SearchResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "image-search",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/centroids/{name}/search": {
      "get": {
        "description": "Search using a loaded centroid as the query vector.",
        "operationId": "search_by_centroid_api_centroids__name__search_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "description": "max results",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 28,
              "description": "max results",
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "offset into the full result set",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "offset into the full result set",
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Retrieval mode. 'centroid' (default) uses the full mean of the seed set. 'sample' picks a random K-subset of the seeds and uses the mean of THAT subset. Each request re-rolls, so refreshing surfaces a different cluster.",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "centroid",
              "description": "Retrieval mode. 'centroid' (default) uses the full mean of the seed set. 'sample' picks a random K-subset of the seeds and uses the mean of THAT subset. Each request re-rolls, so refreshing surfaces a different cluster.",
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "description": "Cluster count (k-means K) for sample mode. Defaults to 10. Only used when mode=sample. Round-75.",
            "in": "query",
            "name": "sample_k",
            "required": false,
            "schema": {
              "default": 10,
              "description": "Cluster count (k-means K) for sample mode. Defaults to 10. Only used when mode=sample. Round-75.",
              "title": "Sample K",
              "type": "integer"
            }
          },
          {
            "description": "Number of cluster centroids to average per request. Defaults to 3. Must satisfy 1 <= n <= sample_k. Only used when mode=sample. Round-75.",
            "in": "query",
            "name": "sample_n",
            "required": false,
            "schema": {
              "default": 3,
              "description": "Number of cluster centroids to average per request. Defaults to 3. Must satisfy 1 <= n <= sample_k. Only used when mode=sample. Round-75.",
              "title": "Sample N",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search By Centroid"
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "api_search_api_search_get",
        "parameters": [
          {
            "description": "text query",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "text query",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "max results",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 28,
              "description": "max results",
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "offset into the full result set",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "offset into the full result set",
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "result view: 'grid' or 'feed'",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "grid",
              "description": "result view: 'grid' or 'feed'",
              "title": "View",
              "type": "string"
            }
          },
          {
            "description": "restrict results to favourites",
            "in": "query",
            "name": "favorites",
            "required": false,
            "schema": {
              "default": false,
              "description": "restrict results to favourites",
              "title": "Favorites",
              "type": "boolean"
            }
          },
          {
            "description": "apply MMR diversity re-ranking",
            "in": "query",
            "name": "diverse",
            "required": false,
            "schema": {
              "default": false,
              "description": "apply MMR diversity re-ranking",
              "title": "Diverse",
              "type": "boolean"
            }
          },
          {
            "description": "Diversity strength: off, low, balanced, or high",
            "in": "query",
            "name": "diversity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Diversity strength: off, low, balanced, or high",
              "title": "Diversity"
            }
          },
          {
            "description": "Diversity candidate depth: auto, 500, 1000, 2000, 5000, or 10000",
            "in": "query",
            "name": "diversity_depth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Diversity candidate depth: auto, 500, 1000, 2000, 5000, or 10000",
              "title": "Diversity Depth"
            }
          },
          {
            "description": "Surprise Me \u2014 random sample from deep pool",
            "in": "query",
            "name": "surprise",
            "required": false,
            "schema": {
              "default": false,
              "description": "Surprise Me \u2014 random sample from deep pool",
              "title": "Surprise",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Search"
      }
    },
...
  }
}